+++ NewPointe/ProfileManager/Structures/AutoJoinProfile.cs
@@ -14,11 +14,11 @@ namespace NewPointe.ProfileManager.Structures
         public string created_at { get; set; }
         public int[] device_groups { get; set; }
         public bool has_complete_data { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public bool is_wildcard { get; set; }
         public string name { get; set; }
         public string reg_challenge { get; set; }
-        public int temporary_id { get; set; }
+        public long temporary_id { get; set; }
         public string updated_at { get; set; }
         public AutoJoinProfileUsageLogItem[] usage_log { get; set; }
     }
